feat(nvim): add nvim-treesitter plugin and configure it
Add nvim-treesitter plugin to packer and configure it in ts.lua file. The
plugin is set up to install parsers for c, lua, vim, vimdoc, and query
languages. It is also set up to automatically install missing parsers
when entering buffer. The highlight feature is enabled and configured
to disable highlighting for c and rust languages and to disable slow
treesitter highlight for large files. Additionally, the plugin is
configured to run `:h syntax` and tree-sitter at the same time.