feat(ft_norm): add script to normalize audio volume using ffmpeg
The `ft_norm` script is added to the `bin` directory. This script is used to normalize the volume of an audio file using ffmpeg. It takes one argument, which is the file to be converted.

The script ensures that the correct number of arguments is provided. If the number of arguments is not equal to 1, it displays a usage message and exits with a status of 1.

The script defines a local variable `filter` which contains the ffmpeg filter to be applied to the audio. The filter normalizes the volume to 30dB.

The script then determines the input file by using the `ls` command to list all files matching the provided argument. The output file is derived from the input file by removing the file extension and appending "_out.mp3".

Finally, the script uses ffmpeg to convert the input file to an output file. It uses the libmp3lame codec,