feat(ft_split): add script to split audio file into chunks
This commit adds a new script called `ft_split` that can be used to split an audio file into smaller chunks. The script takes a filename as a parameter and supports optional parameters for chunk size and overlap duration.

The default values for chunk size and overlap are set to 30 seconds and 5 seconds respectively. The script can be executed with the following command:

```
$ ./ft_split filename [-s|--size chunk_size] [-o|--overlap overlap_duration]
```

The script uses `ffmpeg` to split the audio file into chunks. It calculates the duration of the audio file using `ffprobe` and then creates multiple chunks based on the specified chunk size and overlap duration. Each chunk is saved as an MP3 file with a padded zero index.

Example usage:

```
$ ./ft_split audio.mp3 -s 60 -o 10
```

This will split the `audio.mp