feat(set_dates): add script to set file dates in a directory
This commit adds a new script called `set_dates` to the `bin` directory. The script takes a directory path as an argument and sets the creation date of all files in that directory to the creation date of the directory itself.

The script first checks if a directory path is provided as an argument. If not, it displays an error message and exits with a non-zero status code. Then, it checks if the provided path is a valid directory. If not, it displays an error message and exits with a non-zero status code.

Next, the script retrieves the creation date of the directory using the `stat` command and formats it as `YYYYMMDDHHMM.SS`.

After that, the script sets the creation date of all files in the directory to the retrieved creation date using the `touch` command.

Finally, the script displays a success message indicating that the dates have been set successfully.