chore(karabiner.json): disable COMMAND + TAB shortcut feat(karabiner.json): add key mappings for left_command + left_shift + left_control + left_option + any + h, j, k, l, d, u, w, c, s, f, f1, f2, f3, f10, e, r
chore(karabiner.json): reformat key bindings for 't', 'y', and 'i' keys

The key bindings for the 't', 'y', and 'i' keys in the karabiner.json file were reformatted to improve readability. The modifiers for these keys were moved to separate lines for each modifier, making it easier to understand the key bindings.