chore(nvim): update config for which-key.nvim plugin
The configuration for the which-key.nvim plugin in the nvim.conf-symlink/plugin/packer_compiled.lua file has been updated. The change involves modifying the config value for the plugin. The new config value disables the battery information. This change was made to improve the performance and remove unnecessary information for machines that don't have a battery.