feat(ytd): add script to download embedded videos from web pages to desktop
A new script called "ytd" has been added to the "bin" directory. This script allows users to download embedded videos from any web page directly to their desktop. The script uses the "youtube-dlp" tool, which can be found at https://github.com/yt-dlp/yt-dlp. The script changes the current directory to the user's desktop and then uses "yt-dlp" to download the video specified by the URL passed as an argument to the script.